Backend/payment-service/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
ifaddr==0.1.7
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.26.1
pika==1.2.0
ply==3.11
psycopg2-binary==2.8.6
py-eureka-client==0.10.0
py-zipkin==0.20.2
PyJWT==2.1.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
thriftpy2==0.4.14
urllib3==1.26.5
Werkzeug==2.0.1
